requirements.txt
requests
python-dotenv
anthropic
openai
pillow
moviepy
flask